[fix + refactor] fix on recieving message
- I did the on revieving message wrong last night. I checked the wrong topic name.
- refactored a lot of things
- introduced utils file for the fucntions
- looks cleaner now
