+++ pond/connection.py
@@ -2,105 +2,22 @@ import pygame
 import paho.mqtt.client as mqtt
 import time
 import json
-import random
-import math
 import os
 from PIL import Image
 from fish import Fish
+from utils import *
 #pip install pillow
-
-# MQTT server details
-BROKER = "40.90.169.126"
-PORT = 1883
-USERNAME = "dc24"
-PASSWORD = "kmitl-dc24"
-TOPIC = "fishhaven/stream"
-
-# Define topic for each users
-
-# topic for sending to DC_UNIVERSE group
-DC_UNIVERSE = "user/DC_Universe"
-
-
-# Define the "hello" message
-HELLO_MESSAGE = {
-    "type": "hello",
-    "sender": "Parallel",
-    "timestamp": int(time.time()),
-    "data": {}
-}
-
-SEND_MESSAGE = {
-    "type": "message",
-    "sender": "User",
-    "timestamp": int(time.time()),
-    "data": {
-        "text": "Hello from Pygame button!"
-    }
-}
 
 # Pygame setup
 pygame.init()
-screen_width, screen_height = 1200, 800
 clock = pygame.time.Clock()
 screen = pygame.display.set_mode((screen_width, screen_height))  # Larger screen size (1200x800)
-pygame.display.set_caption("MQTT Message Pond with GIF")
-font = pygame.font.Font(None, 36)
-text_color = (255, 255, 255)  # White text
-bg_color = (26, 148, 49)  # Green background
-button_color = (0, 128, 0)  # Green button
-button_text_color = (255, 255, 255)  # White text for the button
-circle_color = (173, 216, 230)  # Blue circle
+pygame.display.set_caption("Parallel Pond")
 
-# To hold the GIF frames and current animations
+# Pond parameters
 fish_animations = []
 last_update_time = time.time()
-
-# To hold the latest received messages
-received_messages = []
-
-# Load and resize GIF frames
-def load_gif_frames(gif_path):
-    frames = []
-    try:
-        gif = Image.open(gif_path)
-        for frame in range(gif.n_frames):
-            gif.seek(frame)
-            frame_image = gif.convert("RGBA")
-            frame_surface = pygame.image.fromstring(frame_image.tobytes(), frame_image.size, frame_image.mode)
-            frame_surface = pygame.transform.scale(frame_surface, (50, 50))
-            frames.append(frame_surface)
-        return frames
-    except Exception as e:
-        print(f"Unable to load GIF frames: {e}")
-        return []
-
 fish_frames = load_gif_frames("./lib/assets/Parallel.gif")
-# fish_frames = load_gif_frames("./lib/assets/DC_UNIVERSE.gif")
-
-# Callback when the client connects to the broker
-def on_connect(client, userdata, flags, rc):
-    if rc == 0:
-        print("Connected to MQTT broker successfully!")
-        # Publish the "hello" message
-        client.publish(TOPIC, json.dumps(HELLO_MESSAGE))
-        print("Published message:", HELLO_MESSAGE)
-    else:
-        print("Failed to connect, return code %d\n", rc)
-
-# Callback when a message is received from the broker
-def on_message(client, userdata, msg):
-    message = msg.payload.decode()
-    print(f"Message received on topic {msg.topic}: {message}")
-    if msg.topic == "user/DC_Universe":
-        DC_FRAME = load_gif_frames("./lib/assets/DC_Universe.gif")
-        DC_POSITION = generate_random_position(pond_center, pond_radius, fish_frames[0].get_rect())
-        fish_animations.append(Fish(DC_FRAME, DC_POSITION, message.name))
-
-    # Add received message to the list
-    received_messages.append(message)
-    if len(received_messages) > 10:  # Keep only the latest 10 messages
-        received_messages.pop(0)
 
 # Create an MQTT client instance
 client = mqtt.Client()
@@ -122,30 +39,9 @@ client.loop_start()
 client.subscribe(TOPIC)
 client.subscribe("user/Parallel")
 
-def send_fish_to_topicX(topicX, fishName, remainingLifetime):
-    print("Sending message to topic:", topicX)
-    message = { 
-        "name": fishName,
-        "group_name": "Parallel",
-        "lifetime": remainingLifetime
-    } 
-    print(f"Sending message: {message}")
-    client.publish(topicX, json.dumps(message))
-
-# Generate a random position within the circle
-def generate_random_position(center, radius, image_rect):
-    # Random angle in radians
-    angle = random.uniform(0, 2 * math.pi)
-    # Random distance from the center within the radius of the circle
-    distance = random.uniform(0, radius - image_rect.width / 2)
-    # Convert polar coordinates to cartesian coordinates
-    x = center[0] + distance * math.cos(angle)
-    y = center[1] + distance * math.sin(angle)
-    return x, y
 
 # Main loop to keep the script running and display messages
 running = True
-fish_id = 0
 while running:
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
@@ -158,12 +54,11 @@ while running:
             if button_rect_spawn.collidepoint(mouse_x, mouse_y):
                 # Spawn a new fish (GIF) at a random position in the pond
                 fish_position = generate_random_position(pond_center, pond_radius, fish_frames[0].get_rect())
-                fish_animations.append(Fish(fish_frames, fish_position, fish_id))  # Add new fish to the list
-                fish_id += 1  # Increment the fish ID
+                fish_animations.append(Fish(fish_frames, fish_position))  # Add new fish to the list
                 
             # Check if the mouse click is inside the Send Message button area
             if button_rect_send.collidepoint(mouse_x, mouse_y) and fish_animations:
-                send_fish_to_topicX(DC_UNIVERSE, f"Fish{fish_animations[0].name}", fish_animations[0].remainingLifetime)
+                send_fish_to_topicX(client, NETLINK, f"Fish{fish_animations[0].name}", fish_animations[0].remainingLifetime)
                 fish_animations.pop(0)  # Remove the first fish in the list
 
     # Fill the screen with the background color
